Initial commit
Created a small tool to calculate crc32 sums over areas in a .hex or
.bin file and insert the result back into the file.
This is needed for my Zephyr project with a TI CC2340R5 processor, as
the build system cannot currently calculate the crc sums automatically.
Trying to fix it with this tool.
Tried to keep it as general purpose as possible in order to increase
adapdation and have it flexible enough to be also useful for other
devices without having to modify this sourcecode.

Signed-off-by: Fabian Pflug <fabian.pflug@gumulka.eu>